requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.26.0
aws-cdk.aws_lambda_python_alpha
cdk-nag>=2.27.0
